+++ tests/aa.py
@@ -1,4 +1,3 @@
-from pronunciation_mapper import PronunciationMapper
 from pronunciation_mapper.utils import convert_korean_numbers_correctly
 # 테스트 코드
 def test_number_conversion():